demo/requirements.txt
gradio
torch
transformers
Pillow
numpy
